refactor(clear): split app.tsx into rows/lists-screen/editor/swipe modules
app.tsx was past 1000 lines; the seams with the fewest cross-references now
live in their own modules, each owning its visuals outright:

- rows.tsx — the RowSlot pool, the row JSX (icons/edges/strike), and the
  shared motion reset used by park/alloc paths
- lists-screen.tsx — the blue lists canvas factory (rows, count cells,
  credit block) exposing its NodeMirrors for the vertical transitions
- editor.ts — the editing state machine, keyboard handler wiring, row
  shading, and the canvas lift
- swipe.ts — the horizontal swipe gesture with all its transient visuals
  (damped travel, icon rides, strike growth, armed tints); the host commits
  the model mutations

app.tsx (690 lines) keeps the model-to-slot layout pass, scrolling, pull
affordances, screen transitions, and the remaining gestures. No behavior
change: 9 sim journeys and the full 12/12 suite stay green.
