+++ apps/clear/app.tsx
@@ -8,15 +8,16 @@
 // each screen is a full-content-height CANVAS translated inside it — the
 // scroller drives the canvas, absolute 62px rows ride the canvas, and screen
 // switches are vertical canvas moves. The pull-to-create flap folds in real
-// 3D (rotateX under a perspective root); the swipe check/cross are drawn from
-// rotated bars, not glyphs.
+// 3D (rotateX under a perspective root).
 //
-// Rendering strategy: a fixed pool of row nodes mounted once. Data lives in
-// plain arrays (model.ts); after every mutation layout() re-assigns todos to
-// slots. Text and the done look ride per-slot refs; ALL motion (translate/
-// opacity/rotation/gradient colors/height) goes through jump()/animate() on
-// captured NodeMirrors — never through :style objects, whose reactive
-// re-application would clobber in-flight tweens.
+// The module split: rows.tsx owns the row pool + row JSX, lists-screen.tsx
+// the blue lists canvas, swipe.ts the horizontal swipe visuals, editor.ts
+// the edit state + keyboard wiring (keyboard.tsx/kb-layout.ts the keyboard
+// itself). This file owns the model-to-slot layout pass, scrolling, the
+// pull affordances, screen transitions, and the remaining gestures. ALL
+// motion goes through jump()/animate() on captured NodeMirrors — never
+// through :style objects, whose reactive re-application would clobber
+// in-flight tweens.
 
 import { onMounted, shallowRef } from "vue";
 import { Text, View, type NodeMirror } from "@pocketjs/framework/components";
@@ -38,15 +39,7 @@ import {
   MAX_TODOS,
   type Todo,
 } from "./model.ts";
-import {
-  DONE_FROM,
-  DONE_TEXT,
-  DONE_TO,
-  GREEN_FROM,
-  GREEN_TO,
-  listRowColors,
-  todoRowColors,
-} from "./palette.ts";
+import { DONE_FROM, DONE_TEXT, DONE_TO, todoRowColors } from "./palette.ts";
 import {
   OVERSCROLL,
   PINCH_COMMIT,
@@ -56,43 +49,22 @@ import {
   ROW_H,
   SCREEN_H,
   SCREEN_W,
-  SWIPE_COMMIT,
   SWITCH_MS,
   TITLE_FONT_SLOT,
 } from "./metrics.ts";
-import { KB_H, makeKeyboard } from "./keyboard.tsx";
+import { KB_H } from "./keyboard-metrics.ts";
+import { makeSlots, PARKED_Y, renderRow, resetSlotMotion, type RowSlot } from "./rows.tsx";
+import { makeListsScreen } from "./lists-screen.tsx";
+import { makeEditor } from "./editor.ts";
+import { attachSwipeGesture } from "./swipe.ts";
 
 const VIEW_H = SCREEN_H;
 
-/** Off-canvas parking spot for unassigned row slots. */
-const PARKED_Y = 4000;
 /** Where the todos canvas rests while the lists screen owns the display. */
 const TODOS_PARKED_Y = SCREEN_H + 2 * ROW_H;
 
-interface RowSlot {
-  node: NodeMirror | null;
-  front: NodeMirror | null;
-  check: NodeMirror | null;
-  cross: NodeMirror | null;
-  strike: NodeMirror | null;
-  text: ReturnType<typeof shallowRef<string>>;
-  done: ReturnType<typeof shallowRef<boolean>>;
-  lift: ReturnType<typeof shallowRef<number>>;
-  todoId: number;
-  /** Current resting y (display index * ROW_H). */
-  y: number;
-  /** Mid-exit animation: layout() must not repark or reuse it yet. */
-  busy: boolean;
-  gradFrom: string;
-  gradTo: string;
-  /** Measured title width (strike-through line length). */
-  textW: number;
-  textFor: string;
-}
-
 export default () => {
   const lists = seedLists();
-  const LISTS_H = lists.length * ROW_H;
   let activeIndex = 0;
   let screenName: "lists" | "todos" | "switching" = "lists";
   let order: Todo[] = [];
@@ -101,43 +73,18 @@ export default () => {
 
   const list = () => lists[activeIndex];
 
-  let listsCanvas: NodeMirror | null = null;
+  const listsScreen = makeListsScreen(lists);
   let todosCanvas: NodeMirror | null = null;
   let flapNode: NodeMirror | null = null;
   let topSwitchNode: NodeMirror | null = null;
   let previewNode: NodeMirror | null = null;
   let footerNode: NodeMirror | null = null;
-  const listRowNodes: (NodeMirror | null)[] = lists.map(() => null);
 
   const flapText = shallowRef("Pull to Create Item");
   const hasDone = shallowRef(false);
-  const listCounts = lists.map((l) => shallowRef(String(pendingCount(l))));
-  const listEmpty = lists.map((l) => shallowRef(pendingCount(l) === 0));
 
-  const slots: RowSlot[] = Array.from({ length: MAX_TODOS }, () => ({
-    node: null,
-    front: null,
-    check: null,
-    cross: null,
-    strike: null,
-    text: shallowRef(""),
-    done: shallowRef(false),
-    lift: shallowRef(0),
-    todoId: -1,
-    y: PARKED_Y,
-    busy: false,
-    gradFrom: "",
-    gradTo: "",
-    textW: 0,
-    textFor: "",
-  }));
+  const slots = makeSlots(MAX_TODOS);
   const slotByTodo = new Map<number, RowSlot>();
-
-  // -------------------------------------------------------------- editing
-  let editing: Todo | null = null;
-  let editCaret = 0;
-  let editOriginal = "";
-  let editWasNew = false;
 
   // -------------------------------------------------------------- scrolling
   const scroller = createScroller({
@@ -146,7 +93,7 @@ export default () => {
     overscroll: OVERSCROLL,
   });
   const listsScroller = createScroller({
-    max: () => Math.max(0, LISTS_H - VIEW_H),
+    max: () => Math.max(0, listsScreen.height - VIEW_H),
     extent: () => VIEW_H,
     overscroll: OVERSCROLL,
   });
@@ -168,8 +115,7 @@ export default () => {
         slot.todoId = todo.id;
         slot.done.value = todo.done;
         slot.lift.value = 0;
-        if (slot.node) jump(slot.node, "translateX", 0);
-        if (slot.front) jump(slot.front, "translateX", 0);
+        resetSlotMotion(slot);
         slotByTodo.set(todo.id, slot);
         return slot;
       }
@@ -192,8 +138,7 @@ export default () => {
     order = ordered(current);
     contentH = order.length * ROW_H;
     const pending = pendingCount(current);
-    listCounts[activeIndex].value = String(pending);
-    listEmpty[activeIndex].value = pending === 0;
+    listsScreen.refresh(activeIndex);
     hasDone.value = order.length > pending;
     if (todosCanvas) jump(todosCanvas, "height", Math.max(contentH, VIEW_H));
     // The pull-to-clear hint hides below the fold until an overscroll reveals it.
@@ -212,7 +157,7 @@ export default () => {
       const todo = order[index];
       const slot = slotByTodo.get(todo.id) ?? allocSlot(todo);
       slot.done.value = todo.done;
-      if (editing !== todo) slot.text.value = todo.text;
+      if (editor.editing() !== todo) slot.text.value = todo.text;
 
       const y = index * ROW_H;
       if (slot.node) {
@@ -224,7 +169,7 @@ export default () => {
       }
       slot.y = y;
 
-      if (slot.strike && editing !== todo) {
+      if (slot.strike && editor.editing() !== todo) {
         jump(slot.strike, "width", measureTitle(slot, todo.text));
         jump(slot.strike, "scaleX", todo.done ? 1 : 0);
         jump(slot.strike, "bgColor", todo.done ? DONE_TEXT : "#ffffff");
@@ -253,17 +198,19 @@ export default () => {
     slot.gradTo = "";
     slot.lift.value = 0;
     slot.text.value = "";
-    if (slot.node) {
-      jump(slot.node, "translateY", PARKED_Y);
-      jump(slot.node, "translateX", 0);
-    }
-    if (slot.front) {
-      jump(slot.front, "translateX", 0);
-      jump(slot.front, "opacity", 1);
-    }
-    if (slot.check) jump(slot.check, "opacity", 0);
-    if (slot.cross) jump(slot.cross, "opacity", 0);
-    if (slot.strike) jump(slot.strike, "scaleX", 0);
+    if (slot.node) jump(slot.node, "translateY", PARKED_Y);
+    resetSlotMotion(slot);
+  }
+
+  /** A deleted row slides out first; park it once the exit tween lands. */
+  function parkAfterExit(slot: RowSlot): void {
+    after(SWITCH_MS / 1000 + 0.02, () => {
+      slot.busy = false;
+      slot.y = PARKED_Y;
+      slot.text.value = "";
+      if (slot.node) jump(slot.node, "translateY", PARKED_Y);
+      resetSlotMotion(slot);
+    });
   }
 
   /** Display index under a screen y, or -1 outside the rows. */
@@ -273,74 +220,17 @@ export default () => {
   }
 
   // ---------------------------------------------------------------- editor
-  function paintEditRow(): void {
-    if (!editing) return;
-    const slot = slotByTodo.get(editing.id);
-    if (!slot) return;
-    const t = editing.text;
-    slot.text.value = `${t.slice(0, editCaret)}|${t.slice(editCaret)}`;
-  }
-
-  function shadeRows(shaded: boolean): void {
-    const keep = editing ? slotByTodo.get(editing.id) : null;
-    for (const slot of slots) {
-      if (slot.todoId === -1 || slot.busy || !slot.front || slot === keep) continue;
-      animate(slot.front, "opacity", shaded ? 0.15 : 1, { dur: 200, easing: "out" });
-    }
-  }
-
-  function openEditor(todo: Todo, wasNew: boolean): void {
-    editing = todo;
-    editWasNew = wasNew;
-    editOriginal = todo.text;
-    editCaret = todo.text.length;
-    kb.setOpen(true);
-    paintEditRow();
-    shadeRows(true);
-    const index = order.indexOf(todo);
-    const rowBottom = index * ROW_H - scroller.offset() + ROW_H;
-    const liftNeeded = Math.max(0, rowBottom - (SCREEN_H - KB_H));
-    if (todosCanvas) {
-      animate(todosCanvas, "translateY", -scroller.offset() - liftNeeded, { dur: 200, easing: "out" });
-    }
-  }
-
-  function closeEditor(commit: boolean): void {
-    const todo = editing;
-    if (!todo) return;
-    shadeRows(false);
-    editing = null;
-    kb.setOpen(false);
-    if (todosCanvas) {
-      animate(todosCanvas, "translateY", -scroller.offset(), { dur: 200, easing: "out" });
-    }
-    if (commit) {
-      todo.text = todo.text.trim();
-      if (todo.text === "") removeTodo(list(), todo);
-      else report();
-    } else if (editWasNew) {
-      removeTodo(list(), todo);
-    } else {
-      todo.text = editOriginal;
-    }
-    layout(true);
-  }
-
-  const kb = makeKeyboard({
-    onInsert(ch) {
-      if (!editing || editing.text.length >= 40) return;
-      editing.text = editing.text.slice(0, editCaret) + ch + editing.text.slice(editCaret);
-      editCaret += ch.length;
-      paintEditRow();
-    },
-    onBackspace() {
-      if (!editing || editCaret === 0) return;
-      editing.text = editing.text.slice(0, editCaret - 1) + editing.text.slice(editCaret);
-      editCaret -= 1;
-      paintEditRow();
-    },
-    onEnter: () => closeEditor(true),
+  const editor = makeEditor({
+    list,
+    order: () => order,
+    slots,
+    slotFor: (todoId) => slotByTodo.get(todoId),
+    scrollOffset: () => scroller.offset(),
+    canvas: () => todosCanvas,
+    layout,
+    report,
   });
+  const kb = editor.kb;
 
   // ------------------------------------------------------------- creation
   function createAt(index: number): void {
@@ -348,16 +238,17 @@ export default () => {
     if (!todo) return;
     layout(false);
     report();
-    openEditor(todo, true);
+    editor.open(todo, true);
   }
 
   // ------------------------------------------------------------ navigation
   /** Park the lists canvas above the todos so a long pull drags it back in. */
   function parkListsForPulldown(): void {
     listsParked = true;
-    if (listsCanvas) jump(listsCanvas, "translateY", -(LISTS_H + ROW_H));
+    const canvas = listsScreen.canvas();
+    if (canvas) jump(canvas, "translateY", -(listsScreen.height + ROW_H));
     for (let i = 0; i < lists.length; i += 1) {
-      const node = listRowNodes[i];
+      const node = listsScreen.rowNode(i);
       if (node) {
         jump(node, "translateY", i * ROW_H);
         jump(node, "opacity", 1);
@@ -374,12 +265,12 @@ export default () => {
     // The lists clear away vertically: the tapped row rides to the top edge
     // and fades, rows above stack off the top, rows below drop off the bottom.
     for (let i = 0; i < lists.length; i += 1) {
-      const node = listRowNodes[i];
+      const node = listsScreen.rowNode(i);
       if (!node) continue;
       const screenTarget = i <= index ? (i - index) * ROW_H : SCREEN_H + (i - index) * ROW_H;
       animate(node, "translateY", screenTarget + loff, { dur: SWITCH_MS, easing: "out" });
     }
-    const tapped = listRowNodes[index];
+    const tapped = listsScreen.rowNode(index);
     if (tapped) animate(tapped, "opacity", 0, { dur: SWITCH_MS, easing: "out" });
     // The todo stack starts squeezed at the tapped row's screen position and
     // unfolds downward while the canvas rides to the top.
@@ -408,17 +299,14 @@ export default () => {
   }
 
   function goBack(): void {
-    if (editing) closeEditor(false);
+    if (editor.editing()) editor.close(false);
     screenName = "switching";
-    for (let i = 0; i < lists.length; i += 1) {
-      const pending = pendingCount(lists[i]);
-      listCounts[i].value = String(pending);
-      listEmpty[i].value = pending === 0;
-    }
+    listsScreen.refresh();
     // Both canvases travel DOWN: the lists land from above, the todos drop
     // off the bottom.
     listsParked = false;
-    if (listsCanvas) animate(listsCanvas, "translateY", 0, { dur: SWITCH_MS, easing: "out" });
+    const canvas = listsScreen.canvas();
+    if (canvas) animate(canvas, "translateY", 0, { dur: SWITCH_MS, easing: "out" });
     if (todosCanvas) animate(todosCanvas, "translateY", TODOS_PARKED_Y, { dur: SWITCH_MS, easing: "out" });
     scroller.scrollTo(0, { immediate: true });
     paintedOffset = 0;
@@ -435,7 +323,7 @@ export default () => {
 
   // ------------------------------------------------------------- gestures
   const inTodoList = () =>
-    screenName === "todos" && !editing
+    screenName === "todos" && !editor.editing()
       ? { x: 0, y: 0, w: SCREEN_W, h: SCREEN_H }
       : null;
   const inLists = () =>
@@ -481,148 +369,28 @@ export default () => {
   });
 
   // Horizontal pan on a row: complete (right) / delete (left).
-  let swipeSlot: RowSlot | null = null;
-  let swipeTodo: Todo | null = null;
-  let swipeIndex = -1;
-  let swipeArmed = false;
-
-  /** 1:1 up to the commit bound, then damped to a third — the reference feel. */
-  function swipeDisplay(dx: number): number {
-    if (dx > SWIPE_COMMIT) return SWIPE_COMMIT + (dx - SWIPE_COMMIT) / 3;
-    if (dx < -SWIPE_COMMIT) return -SWIPE_COMMIT + (dx + SWIPE_COMMIT) / 3;
-    return dx;
-  }
-
-  /** Past the rightward bound the slider changes its mind: a pending row goes
-   *  green, a done row previews its restored position tint. */
-  function paintArmed(slot: RowSlot, todo: Todo, index: number, armed: boolean): void {
-    if (!slot.front) return;
-    if (armed) {
-      if (todo.done) {
-        slot.done.value = false;
-        const [from, to] = todoRowColors(Math.min(index, 7), 1);
-        jump(slot.front, "gradFrom", from);
-        jump(slot.front, "gradTo", to);
-      } else {
-        jump(slot.front, "gradFrom", GREEN_FROM);
-        jump(slot.front, "gradTo", GREEN_TO);
-      }
-    } else if (todo.done) {
-      slot.done.value = true;
-      jump(slot.front, "gradFrom", DONE_FROM);
-      jump(slot.front, "gradTo", DONE_TO);
-    } else {
-      jump(slot.front, "gradFrom", slot.gradFrom);
-      jump(slot.front, "gradTo", slot.gradTo);
-    }
-  }
-
-  createGesture({
-    axis: "x",
-    region: { rect: inTodoList },
-    onPanStart: (c) => {
-      swipeIndex = rowIndexAt(c.startY);
-      swipeTodo = swipeIndex >= 0 ? order[swipeIndex] : null;
-      swipeSlot = swipeTodo ? (slotByTodo.get(swipeTodo.id) ?? null) : null;
-      swipeArmed = false;
+  attachSwipeGesture({
+    region: inTodoList,
+    rowIndexAt,
+    rowAt: (index) => {
+      const todo = order[index];
+      return todo ? { slot: slotByTodo.get(todo.id) ?? null, todo } : null;
     },
-    onPanMove: (c) => {
-      const slot = swipeSlot;
-      const todo = swipeTodo;
-      if (!slot?.front || !todo) return;
-      const raw = Math.max(-SCREEN_W, Math.min(SCREEN_W, c.dx));
-      const disp = swipeDisplay(raw);
-      jump(slot.front, "translateX", disp);
-      const rightO = Math.max(0, Math.min(1, raw / SWIPE_COMMIT));
-      const leftO = Math.max(0, Math.min(1, -raw / SWIPE_COMMIT));
-      if (slot.check) {
-        jump(slot.check, "opacity", raw > 0 ? (todo.done ? 1 - rightO : rightO) : 0);
-        jump(slot.check, "translateX", Math.max(0, disp - SWIPE_COMMIT));
-      }
-      if (slot.cross) {
-        jump(slot.cross, "opacity", leftO);
-        jump(slot.cross, "translateX", Math.min(0, disp + SWIPE_COMMIT));
-      }
-      if (slot.strike) {
-        jump(slot.strike, "scaleX", todo.done ? 1 - rightO : rightO);
-      }
-      const armed = raw >= SWIPE_COMMIT;
-      if (armed !== swipeArmed) {
-        swipeArmed = armed;
-        paintArmed(slot, todo, swipeIndex, armed);
-      }
+    complete(todo) {
+      setDone(list(), todo, !todo.done);
+      layout(true);
+      report();
     },
-    onPanEnd: (c) => {
-      const slot = swipeSlot;
-      const todo = swipeTodo;
-      swipeSlot = null;
-      swipeTodo = null;
-      if (!slot?.front || !todo) return;
-      if (slot.check) animate(slot.check, "opacity", 0, { dur: 160, easing: "out" });
-      if (slot.cross) animate(slot.cross, "opacity", 0, { dur: 160, easing: "out" });
-      if (c.dx >= SWIPE_COMMIT) {
-        animate(slot.front, "translateX", 0, { dur: 180, easing: "out" });
-        swipeArmed = false;
-        setDone(list(), todo, !todo.done);
-        layout(true);
-        report();
-        return;
-      }
-      if (c.dx <= -SWIPE_COMMIT) {
-        slot.busy = true;
-        parkAfterExit(slot);
-        if (slot.node) {
-          animate(slot.node, "translateX", -(SCREEN_W + ROW_H), { dur: SWITCH_MS, easing: "out" });
-        }
-        removeTodo(list(), todo);
-        slotByTodo.delete(todo.id);
-        slot.todoId = -1;
-        layout(true);
-        report();
-        return;
-      }
-      animate(slot.front, "translateX", 0, { dur: 160, easing: "out" });
-      if (slot.strike) jump(slot.strike, "scaleX", todo.done ? 1 : 0);
-      if (swipeArmed) {
-        swipeArmed = false;
-        paintArmed(slot, todo, swipeIndex, false);
-      }
-    },
-    onCancel: () => {
-      const slot = swipeSlot;
-      const todo = swipeTodo;
-      swipeSlot = null;
-      swipeTodo = null;
-      if (!slot?.front || !todo) return;
-      animate(slot.front, "translateX", 0, { dur: 160, easing: "out" });
-      if (slot.check) animate(slot.check, "opacity", 0, { dur: 160, easing: "out" });
-      if (slot.cross) animate(slot.cross, "opacity", 0, { dur: 160, easing: "out" });
-      if (slot.strike) jump(slot.strike, "scaleX", todo.done ? 1 : 0);
-      if (swipeArmed) {
-        swipeArmed = false;
-        paintArmed(slot, todo, swipeIndex, false);
-      }
+    remove(todo, slot) {
+      slot.busy = true;
+      parkAfterExit(slot);
+      removeTodo(list(), todo);
+      slotByTodo.delete(todo.id);
+      slot.todoId = -1;
+      layout(true);
+      report();
     },
   });
-
-  function parkAfterExit(slot: RowSlot): void {
-    after(SWITCH_MS / 1000 + 0.02, () => {
-      slot.busy = false;
-      slot.y = PARKED_Y;
-      slot.text.value = "";
-      if (slot.node) {
-        jump(slot.node, "translateY", PARKED_Y);
-        jump(slot.node, "translateX", 0);
-      }
-      if (slot.front) {
-        jump(slot.front, "translateX", 0);
-        jump(slot.front, "opacity", 1);
-      }
-      if (slot.check) jump(slot.check, "opacity", 0);
-      if (slot.cross) jump(slot.cross, "opacity", 0);
-      if (slot.strike) jump(slot.strike, "scaleX", 0);
-    });
-  }
 
   // Long-press: pick a pending row up and reorder it.
   let dragSlot: RowSlot | null = null;
@@ -702,7 +470,7 @@ export default () => {
     onTap: (c) => {
       const index = rowIndexAt(c.y);
       if (index >= 0 && !order[index].done) {
-        openEditor(order[index], false);
+        editor.open(order[index], false);
         return;
       }
       createAt(pendingCount(list()));
@@ -780,11 +548,11 @@ export default () => {
   createGesture({
     region: {
       rect: () =>
-        screenName === "todos" && editing
+        screenName === "todos" && editor.editing()
           ? { x: 0, y: 0, w: SCREEN_W, h: SCREEN_H - KB_H }
           : null,
     },
-    onTap: () => closeEditor(true),
+    onTap: () => editor.close(true),
   });
 
   // The keyboard claims its panel outright (registered last = top priority).
@@ -802,6 +570,7 @@ export default () => {
    *  rides down into view. */
   function paintPull(pull: number): void {
     if (!flapNode) return;
+    const listsCanvas = listsScreen.canvas();
     if (pull <= 0) {
       jump(flapNode, "opacity", 0);
       if (topSwitchNode) jump(topSwitchNode, "opacity", 0);
@@ -812,7 +581,7 @@ export default () => {
       jump(flapNode, "opacity", 0);
       if (topSwitchNode) jump(topSwitchNode, "opacity", 1);
       listsParked = false;
-      if (listsCanvas) jump(listsCanvas, "translateY", pull - PULL_BACK - LISTS_H);
+      if (listsCanvas) jump(listsCanvas, "translateY", pull - PULL_BACK - listsScreen.height);
       return;
     }
     if (topSwitchNode) jump(topSwitchNode, "opacity", 0);
@@ -829,7 +598,7 @@ export default () => {
     if (screenName === "todos") {
       scroller.step();
       const off = scroller.offset();
-      if (off !== paintedOffset && todosCanvas && !editing) {
+      if (off !== paintedOffset && todosCanvas && !editor.editing()) {
         paintedOffset = off;
         jump(todosCanvas, "translateY", -off);
       }
@@ -841,6 +610,7 @@ export default () => {
     } else if (screenName === "lists") {
       listsScroller.step();
       const loff = listsScroller.offset();
+      const listsCanvas = listsScreen.canvas();
       if (loff !== paintedListsOffset && listsCanvas) {
         paintedListsOffset = loff;
         jump(listsCanvas, "translateY", -loff);
@@ -860,93 +630,6 @@ export default () => {
   });
 
   // -------------------------------------------------------------- render
-  function renderRow(slot: RowSlot) {
-    return (
-      <View
-        nodeRef={(node) => {
-          slot.node = node ?? null;
-        }}
-        class="absolute left-0 right-0 top-0 h-[62]"
-        style={{ zIndex: slot.lift.value, shadow: slot.lift.value ? 3 : 0 }}
-      >
-        <View
-          nodeRef={(node) => {
-            slot.check = node ?? null;
-          }}
-          class="absolute left-0 top-0"
-          style={{ width: ROW_H, height: ROW_H, opacity: 0 }}
-        >
-          {/* Both strokes overshoot the valley center (23, 39.5) by half a
-              thickness so the rotated bars overlap into one clean joint. */}
-          <View class="absolute bg-[#ffffff]" style={{ insetL: 9.3, insetT: 31.3, width: 19, height: 7, rotate: 45 }} />
-          <View class="absolute bg-[#ffffff]" style={{ insetL: 15.3, insetT: 25.8, width: 35, height: 7, rotate: -45 }} />
-        </View>
-        <View
-          nodeRef={(node) => {
-            slot.cross = node ?? null;
-          }}
-          class="absolute right-0 top-0"
-          style={{ width: ROW_H, height: ROW_H, opacity: 0 }}
-        >
-          <View class="absolute bg-[#eb0017]" style={{ insetL: 15, insetT: 28, width: 32, height: 7, rotate: 45 }} />
-          <View class="absolute bg-[#eb0017]" style={{ insetL: 15, insetT: 28, width: 32, height: 7, rotate: -45 }} />
-        </View>
-        <View
-          nodeRef={(node) => {
-            slot.front = node ?? null;
-          }}
-          class="absolute inset-0 bg-gradient-to-b from-[#f50018] to-[#e00016]"
-        >
-          <View class="absolute left-0 right-0 top-0 bg-[#ffffff12]" style={{ height: 1 }} />
-          <View class="absolute left-0 right-0 bottom-0 bg-[#0000001a]" style={{ height: 1 }} />
-          <View class="absolute inset-0 flex-row items-center pl-3">
-            <Text class={slot.done.value ? "text-xl font-bold text-[#666666]" : "text-xl font-bold text-white"}>
-              {slot.text.value}
-            </Text>
-          </View>
-          <View
-            nodeRef={(node) => {
-              slot.strike = node ?? null;
-            }}
-            class="absolute bg-[#ffffff]"
-            style={{ insetL: 12, insetT: 30, height: 2, width: 0, originX: -0.5, scaleX: 0 }}
-          />
-        </View>
-      </View>
-    );
-  }
-
-  function renderListRow(title: string, i: number) {
-    const [from, to] = listRowColors(i, lists.length);
-    return (
-      <View
-        nodeRef={(node) => {
-          listRowNodes[i] = node ?? null;
-        }}
-        class="absolute left-0 right-0 top-0 h-[62]"
-        style={{ translateY: i * ROW_H }}
-      >
-        <View
-          class="absolute inset-0 bg-gradient-to-b from-[#1780f7] to-[#1780f7]"
-          style={{ gradFrom: from, gradTo: to }}
-        >
-          <View class="absolute left-0 right-0 top-0 bg-[#ffffff12]" style={{ height: 1 }} />
-          <View class="absolute left-0 right-0 bottom-0 bg-[#0000001a]" style={{ height: 1 }} />
-          <View class="absolute inset-0 flex-row items-center pl-3">
-            <Text class={listEmpty[i].value ? "text-xl font-bold text-[#ffffff80]" : "text-xl font-bold text-white"}>
-              {title}
-            </Text>
-          </View>
-          <View class="absolute right-0 top-0 bottom-0 items-center justify-center bg-[#ffffff26]" style={{ width: ROW_H }}>
-            <Text class={listEmpty[i].value ? "text-xl font-bold text-[#ffffff80]" : "text-xl font-bold text-white"}>
-              {listCounts[i].value}
-            </Text>
-          </View>
-        </View>
-      </View>
-    );
-  }
-
   return (
     <View class="w-full h-full bg-[#000000] overflow-hidden">
       <View
@@ -999,22 +682,7 @@ export default () => {
         {slots.map((slot) => renderRow(slot))}
       </View>
 
-      <View
-        nodeRef={(node) => {
-          listsCanvas = node ?? null;
-        }}
-        class="absolute left-0 right-0 top-0"
-        style={{ height: LISTS_H, zIndex: 1 }}
-      >
-        <View
-          class="absolute left-0 right-0 flex-col items-center justify-center"
-          style={{ translateY: -2.5 * ROW_H, height: ROW_H }}
-        >
-          <Text class="text-sm text-[#ffffff40] text-center">Made with PocketJS + Vue Vapor</Text>
-          <Text class="text-sm text-[#ffffff40] text-center">Original by Evan You, iOS app by Realmac</Text>
-        </View>
-        {lists.map((l, i) => renderListRow(l.title, i))}
-      </View>
+      {listsScreen.view}
 
       {kb.view}
     </View>
